moving flow functions to services
